+++ package/src/vite/dev.ts
@@ -17,6 +17,47 @@ type CreateMiddlewareOptions = {
     server: Server;
 };
 
+const createRequestHeaders = (headers: HTTP.IncomingHttpHeaders): Headers => {
+    const result: Headers = new Headers();
+
+    const entries: [
+        string,
+        string | string[] | undefined,
+    ][] = Object.entries(headers);
+
+    for (let i: number = 0; i < entries.length; i++) {
+        const entry:
+            | [
+                  string,
+                  string | string[] | undefined,
+              ]
+            | undefined = entries[i];
+
+        if (entry === void 0) continue;
+
+        const [key, value] = entry;
+
+        // ignore HTTP/2 pseudo-headers
+        if (key.startsWith(":")) continue;
+
+        if (value === void 0) continue;
+
+        if (Array.isArray(value)) {
+            for (let j: number = 0; j < value.length; j++) {
+                const vl: string | undefined = value[j];
+
+                if (vl === void 0) continue;
+
+                result.append(key, vl);
+            }
+        } else {
+            result.set(key, value);
+        }
+    }
+
+    return result;
+};
+
 const createMiddleware = ({ vite, server }: CreateMiddlewareOptions) => {
     return async (
         req: Connect.IncomingMessage,
@@ -42,7 +83,7 @@ const createMiddleware = ({ vite, server }: CreateMiddlewareOptions) => {
 
         const request: Request = new Request(url, {
             method: req.method,
-            headers: req.headers,
+            headers: createRequestHeaders(req.headers),
             body,
             duplex: "half",
         } as RequestInit);