+++ package/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### What's Changed
 
+- fix dev server request header forwarding with https plugins
 - update package metadata
 
 ## 0.2.2 (2026-03-14)